requirements.txt
keyauth
colorama
fade
pycryptodome
requests
pywin32